Refactory: 인증 로직 리팩터링
- AuthenticationConverter 추상화
- AuthenticationInterceptor 추상화
