+++ src/main/java/nextstep/auth/authentication/AuthenticationConverter.java
@@ -1,0 +1,9 @@
+package nextstep.auth.authentication;
+
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface AuthenticationConverter {
+	AuthenticationToken convert(HttpServletRequest request) throws IOException;
+}
