+++ src/main/java/nextstep/auth/authentication/TokenAuthenticationConverter.java
@@ -1,0 +1,23 @@
+package nextstep.auth.authentication;
+
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.json.JSONObject;
+
+public class TokenAuthenticationConverter implements AuthenticationConverter {
+	private static final String EMAIL = "email";
+	private static final String PASSWORD = "password";
+
+	@Override
+	public AuthenticationToken convert(HttpServletRequest request) throws IOException {
+		String stringJson = request.getReader()
+			.lines()
+			.findFirst()
+			.orElseThrow(RuntimeException::new);
+		JSONObject jsonObject = new JSONObject(stringJson);
+
+		return new AuthenticationToken(jsonObject.getString(EMAIL), jsonObject.getString(PASSWORD));
+	}
+}
