+++ src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor.java
@@ -15,21 +15,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
-    private static final String EMAIL = "email";
-    private static final String PASSWORD = "password";
-
-    private CustomUserDetailsService customUserDetailsService;
+public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
     private JwtTokenProvider jwtTokenProvider;
+    private TokenAuthenticationConverter tokenAuthenticationConverter;
 
     public TokenAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService, JwtTokenProvider jwtTokenProvider) {
-        this.customUserDetailsService = customUserDetailsService;
+        super(customUserDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
+        this.tokenAuthenticationConverter = new TokenAuthenticationConverter();
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = convert(request);
+        AuthenticationToken authenticationToken = tokenAuthenticationConverter.convert(request);
         Authentication authentication = authenticate(authenticationToken);
 
         String payload = new ObjectMapper().writeValueAsString(authentication.getPrincipal());
@@ -43,33 +41,5 @@ public class TokenAuthenticationInterceptor implements HandlerInterceptor {
         response.getOutputStream().print(responseToClient);
 
         return false;
-    }
-
-    public AuthenticationToken convert(HttpServletRequest request) throws IOException {
-        String stringJson = request.getReader()
-            .lines()
-            .findFirst()
-            .orElseThrow(RuntimeException::new);
-        JSONObject jsonObject = new JSONObject(stringJson);
-
-        return new AuthenticationToken(jsonObject.getString(EMAIL), jsonObject.getString(PASSWORD));
-    }
-
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        String principal = authenticationToken.getPrincipal();
-        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, authenticationToken);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new AuthenticationException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new AuthenticationException();
-        }
     }
 }
