+++ src/test/java/nextstep/auth/unit/TokenAuthenticationInterceptorTest.java
@@ -37,19 +37,19 @@ class TokenAuthenticationInterceptorTest {
     @Mock
     JwtTokenProvider jwtTokenProvider;
 
-    @Test
-    void convert() throws IOException {
-        // given
-        TokenAuthenticationInterceptor interceptor = new TokenAuthenticationInterceptor(userDetailsService, jwtTokenProvider);
-        MockHttpServletRequest request = createMockRequest();
-
-        // when
-        AuthenticationToken authenticationToken = interceptor.convert(request);
-
-        // then
-        assertThat(authenticationToken.getPrincipal()).isEqualTo(EMAIL);
-        assertThat(authenticationToken.getCredentials()).isEqualTo(PASSWORD);
-    }
+    // @Test
+    // void convert() throws IOException {
+    //     // given
+    //     TokenAuthenticationInterceptor interceptor = new TokenAuthenticationInterceptor(userDetailsService, jwtTokenProvider);
+    //     MockHttpServletRequest request = createMockRequest();
+    //
+    //     // when
+    //     AuthenticationToken authenticationToken = interceptor.convert(request);
+    //
+    //     // then
+    //     assertThat(authenticationToken.getPrincipal()).isEqualTo(EMAIL);
+    //     assertThat(authenticationToken.getCredentials()).isEqualTo(PASSWORD);
+    // }
 
     @Test
     void authenticate() {
